+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # docs
 
+## 0.1.5
+
+### Patch Changes
+
+- Adding a new roundedWhite variant to the Select component
+  - @global-x-etfs/gx-components@3.2.5
+
 ## 0.1.4
 
 ### Patch Changes